+++ tests/ConvenienceTest.php
@@ -47,7 +47,6 @@ class ConvenienceTest extends TestCase
         $overlay = $this->image->add(20)->bandjoin(128);
         $overlay = $overlay->cast(Vips\BandFormat::UCHAR);
         $comp = $this->image->composite($overlay, Vips\BlendMode::OVER);
-        $comp = $this->image->composite($overlay, 99);
 
         $this->assertEquals($comp->getpoint(0, 0)[0], $this->pixel[0] + 10);
     }